Rename the master.py to dist_master.py
